refactor(vc): one slot-outlet collector — export it from @core/visualComponents
The editor store's vcSlotReconcile.ts carried a verbatim copy of slotSync's
unexported extractSlotNamesFromVCTree (its doc comment even said so). Slot
pairing is core VC semantics — two copies of the DFS collector would drift
the editor's no-op-reconcile check away from what syncSlotInstances actually
reads.

The canonical collector is now `collectSlotOutletNames(tree)` (generalized
to the tree shape both callers want), exported from @core/visualComponents;
the slice copy is deleted and site/helpers.ts imports the core one.
