+++ src/admin/pages/site/store/slices/site/helpers.ts
@@ -14,7 +14,8 @@ import type { FrameworkColorToken } from '@core/framework-schema'
 import type { NodeTree, PageNode, StyleRule, SiteDocument } from '@core/page-tree'
 import { addPage, createNode, reconcileSiteExplorerInPlace, reindexNodeParents } from '@core/page-tree'
 import type { VisualComponent } from '@core/visualComponents'
-import { syncAllVCRefSlotInstances, allTreeNodeMaps, collectVCSlotOutletNames } from '../vcSlotReconcile'
+import { syncAllVCRefSlotInstances, allTreeNodeMaps } from '../vcSlotReconcile'
+import { collectSlotOutletNames } from '@core/visualComponents'
 import { create } from 'mutative'
 import type { Draft, Patches } from 'mutative'
 import type { ImportFragment } from '@core/htmlImport'
@@ -335,12 +336,12 @@ export function buildSiteHelpers(
     // `get()` is the frozen pre-mutation state — the VC is guaranteed present
     // there because the draft (where it was just found) mirrors it.
     const frozenVc = get().site?.visualComponents.find((v) => v.id === vc.id)
-    const outletNamesBefore = frozenVc ? collectVCSlotOutletNames(frozenVc.tree) : null
+    const outletNamesBefore = frozenVc ? collectSlotOutletNames(frozenVc.tree) : null
 
     const result = fn(tree)
     if (result === false) return false
 
-    const outletNamesAfter = collectVCSlotOutletNames(vc.tree)
+    const outletNamesAfter = collectSlotOutletNames(vc.tree)
     if (outletNamesBefore === null || !stringArraysEqual(outletNamesBefore, outletNamesAfter)) {
       syncAllVCRefSlotInstances(allTreeNodeMaps(draft.site!), vc.id, vc)
     }
