+++ src/core/visualComponents/index.ts
@@ -22,7 +22,7 @@ export { getReferencedComponentIds, wouldCreateCycle } from './recursionGuard'
 export { forEachVCRef, collectVCRefs } from './vcRefs'
 export type { VCRef } from './vcRefs'
 
-export { syncSlotInstances, applySlotSyncResult } from './slotSync'
+export { collectSlotOutletNames, syncSlotInstances, applySlotSyncResult } from './slotSync'
 
 export { previewVCDeletion } from './deletionImpact'
 export type { VCDeletionImpact, VCRefUsage } from './deletionImpact'
